pyproject.toml
[project]
name = "mcp-webanalyzer"
version = "0.1.0"
description = "Enterprise-grade MCP server for web page analysis and extraction"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "urllib3>=2.0.0",
    "html2text>=2020.1.16",
    "readability-lxml>=0.8.1",
    "newspaper3k>=0.2.8",
    "selenium>=4.15.0",
    "requests>=2.31.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "flower>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "slowapi>=0.1.9",
    "aiofiles>=23.2.1",
    "jinja2>=3.1.2",
    "openai>=1.0.0",
    "anthropic>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
mcp-webanalyzer = "mcp_webanalyzer.server:main"
mcp-webanalyzer-api = "mcp_webanalyzer.api_server:main"
mcp-webanalyzer-worker = "mcp_webanalyzer.worker:main"
mcp-webanalyzer-client = "mcp_webanalyzer.mcp_client:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_webanalyzer"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--cov=mcp_webanalyzer --cov-report=term-missing"